# docs/openapi.yaml
openapi: 3.0.3
info:
  title: macauth API
  description: |
    Authentication microservice designed for **service-to-service** communication.

    All tokens (access & refresh) are exchanged via request/response body — no cookies are used.
    This allows the consuming service to decide how to deliver tokens to end users.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:2800/macauth/api/v1
    description: Local development server
...
components:
...
  schemas:
...
    InitiateResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    ConfirmResetRequest:
      type: object
      required: [resetToken, newPassword]
      properties:
        resetToken:
          type: string
          description: Raw token received by the user.
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 64
          example: NewSecurePassword123!

    # -- Permissions Requests --
...
    InitiateResetResponse:
      type: object
      properties:
        resetToken:
          type: string
          description: Raw token to be sent to the user via email.
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a7b8c9d0e1f2
        email:
          type: string
          format: email
          example: user@example.com
...
    ErrorResponse:
      type: object
      description: Standard error response returned by all endpoints.
      properties:
        code:
          type: integer
          description: HTTP status code.
          example: 400
        message:
          type: string
          description: Human-readable error message.
          example: Invalid password
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred.

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /reset/initiate-reset:
    post:
      tags: [Reset]
      summary: Initiate Password Reset
      operationId: initiateReset
      description: |
        Initiate a password reset for a user. Generates a short-lived reset token
        which is returned to the calling service. The calling service is responsible
        for sending the reset link to the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateResetRequest"
      responses:
        "200":
          description: Reset token successfully generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiateResetResponse"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User with this email does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reset/confirm-reset:
    post:
      tags: [Reset]
      summary: Confirm Password Reset
      operationId: confirmReset
      description: |
        Confirm a password reset using the generated reset token and update the user's password.
        Invalidates all active user sessions (refresh tokens).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmResetRequest"
      responses:
        "200":
          description: Password successfully updated and all active sessions invalidated.
        "400":
          description: Invalid token, token used, expired token, or invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # =========================================================================
  # Permissions API
  # =========================================================================
...